pyproject.toml
[project]
name = "claimtrackr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-pinecone==0.2.13",
    "langchain-text-splitters>=1.1.1",
    "pdfplumber>=0.11.9",
    "pinecone==7.3.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
